pyproject.toml
[project]
name = "qvis"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pysimdjson@git+https://github.com/birneee/pysimdjson#f023fb181b5ebea53514d02536734b2ceb5ffd0d",
    "matplotlib>=3.7.1",
    "pandas>=2.0.2",
    "pyqt5>=5.15.9",
    "numpy>=1.22"
]
